requirements.txt
ultralytics
plotly
plotly-express
Pillow
paddleocr
paddlepaddle
opencv-python
opencv-python-headless
mmdet
mmocr
mycolorpy
easyocr
sentence-transformers
uvicorn
ctransformers
fastapi
ipykernel
langchain
python-box
streamlit

